refactor(test): factor common LOAKit setup
