+++ packages/orchestration/test/exos/local-orchestration-account-kit.test.ts
@@ -3,56 +3,23 @@ import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 import { AmountMath } from '@agoric/ertp';
 import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
 import { heapVowE as E } from '@agoric/vow/vat.js';
-import { prepareRecorderKitMakers } from '@agoric/zoe/src/contractSupport/recorder.js';
-import { Far } from '@endo/far';
-import { TimeMath } from '@agoric/time';
-import { prepareLocalOrchestrationAccountKit } from '../../src/exos/local-orchestration-account.js';
+import { TargetApp } from '@agoric/vats/src/bridge-target.js';
 import { ChainAddress } from '../../src/orchestration-api.js';
-import { makeChainHub } from '../../src/exos/chain-hub.js';
 import { NANOSECONDS_PER_SECOND } from '../../src/utils/time.js';
 import { commonSetup } from '../supports.js';
 import { UNBOND_PERIOD_SECONDS } from '../ibc-mocks.js';
 import { maxClockSkew } from '../../src/utils/cosmos.js';
+import { prepareMakeTestLOAKit } from './make-test-loa-kit.js';
 
 test('deposit, withdraw', async t => {
-  const { bootstrap, brands, utils } = await commonSetup(t);
+  const common = await commonSetup(t);
+  const makeTestLOAKit = prepareMakeTestLOAKit(t, common.bootstrap);
+  const account = await makeTestLOAKit();
 
-  const { bld: stake } = brands;
-
-  const { timer, localchain, marshaller, rootZone, storage, vowTools } =
-    bootstrap;
-
-  t.log('chainInfo mocked via `prepareMockChainInfo` until #8879');
-
-  t.log('exo setup - prepareLocalChainAccountKit');
-  const { makeRecorderKit } = prepareRecorderKitMakers(
-    rootZone.mapStore('recorder'),
-    marshaller,
-  );
-  const makeLocalOrchestrationAccountKit = prepareLocalOrchestrationAccountKit(
-    rootZone,
-    makeRecorderKit,
-    // @ts-expect-error mocked zcf. use `stake-bld.contract.test.ts` to test LCA with offer
-    Far('MockZCF', {}),
-    timer,
-    vowTools,
-    makeChainHub(bootstrap.agoricNames, vowTools),
-  );
-
-  t.log('request account from vat-localchain');
-  const lca = await E(localchain).makeAccount();
-  const address = await E(lca).getAddress();
-
-  t.log('make a LocalChainAccountKit');
-  const { holder: account } = makeLocalOrchestrationAccountKit({
-    account: lca,
-    address: harden({
-      value: address,
-      chainId: 'agoric-n',
-      encoding: 'bech32',
-    }),
-    storageNode: storage.rootNode.makeChildNode('lcaKit'),
-  });
+  const {
+    brands: { bld: stake },
+    utils,
+  } = common;
 
   const oneHundredStakePmt = await utils.pourPayment(stake.units(100));
 
@@ -88,42 +55,15 @@ test('deposit, withdraw', async t => {
 });
 
 test('delegate, undelegate', async t => {
-  const { bootstrap, brands, utils } = await commonSetup(t);
+  const common = await commonSetup(t);
+  const makeTestLOAKit = prepareMakeTestLOAKit(t, common.bootstrap);
+  const account = await makeTestLOAKit();
 
-  const { bld } = brands;
-
-  const { timer, localchain, marshaller, rootZone, storage, vowTools } =
-    bootstrap;
-
-  t.log('exo setup - prepareLocalChainAccountKit');
-  const { makeRecorderKit } = prepareRecorderKitMakers(
-    rootZone.mapStore('recorder'),
-    marshaller,
-  );
-  const makeLocalOrchestrationAccountKit = prepareLocalOrchestrationAccountKit(
-    rootZone,
-    makeRecorderKit,
-    // @ts-expect-error mocked zcf. use `stake-bld.contract.test.ts` to test LCA with offer
-    Far('MockZCF', {}),
-    timer,
-    vowTools,
-    makeChainHub(bootstrap.agoricNames, vowTools),
-  );
-
-  t.log('request account from vat-localchain');
-  const lca = await E(localchain).makeAccount();
-  const address = await E(lca).getAddress();
-
-  t.log('make a LocalChainAccountKit');
-  const { holder: account } = makeLocalOrchestrationAccountKit({
-    account: lca,
-    address: harden({
-      value: address,
-      chainId: 'agoric-n',
-      encoding: 'bech32',
-    }),
-    storageNode: storage.rootNode.makeChildNode('lcaKit'),
-  });
+  const {
+    bootstrap: { timer },
+    brands: { bld },
+    utils,
+  } = common;
 
   await E(account).deposit(await utils.pourPayment(bld.units(100)));
 
@@ -151,42 +91,14 @@ test('delegate, undelegate', async t => {
 });
 
 test('transfer', async t => {
-  const { bootstrap, brands, utils } = await commonSetup(t);
+  const common = await commonSetup(t);
+  const makeTestLOAKit = prepareMakeTestLOAKit(t, common.bootstrap);
+  const account = await makeTestLOAKit();
 
-  const { bld: stake } = brands;
-
-  const { timer, localchain, marshaller, rootZone, storage, vowTools } =
-    bootstrap;
-
-  t.log('exo setup - prepareLocalChainAccountKit');
-  const { makeRecorderKit } = prepareRecorderKitMakers(
-    rootZone.mapStore('recorder'),
-    marshaller,
-  );
-  const makeLocalOrchestrationAccountKit = prepareLocalOrchestrationAccountKit(
-    rootZone,
-    makeRecorderKit,
-    // @ts-expect-error mocked zcf. use `stake-bld.contract.test.ts` to test LCA with offer
-    Far('MockZCF', {}),
-    timer,
-    vowTools,
-    makeChainHub(bootstrap.agoricNames, vowTools),
-  );
-
-  t.log('request account from vat-localchain');
-  const lca = await E(localchain).makeAccount();
-  const address = await E(lca).getAddress();
-
-  t.log('make a LocalChainAccountKit');
-  const { holder: account } = makeLocalOrchestrationAccountKit({
-    account: lca,
-    address: harden({
-      value: address,
-      chainId: 'agoric-n',
-      encoding: 'bech32',
-    }),
-    storageNode: storage.rootNode.makeChildNode('lcaKit'),
-  });
+  const {
+    brands: { bld: stake },
+    utils,
+  } = common;
 
   t.truthy(account, 'account is returned');
